Removed a lot of the overriding functions
