+++ adafruit_featherwing/neopixel_featherwing.py
@@ -33,9 +33,9 @@ __repo__ = "https://github.com/adafruit/Adafruit_CircuitPython_FeatherWing.git"
 
 import board
 import neopixel
-from adafruit_featherwing.pixelmatrix_featherwing import PixelMatrixFeatherWing
+from adafruit_featherwing.pixelmatrix import PixelMatrix
 
-class NeoPixelFeatherWing(PixelMatrixFeatherWing):
+class NeoPixelFeatherWing(PixelMatrix):
     """Class representing a `NeoPixel FeatherWing
        <https://www.adafruit.com/product/2945>`_.
 
@@ -51,118 +51,6 @@ class NeoPixelFeatherWing(PixelMatrixFeatherWing):
         self._matrix = neopixel.NeoPixel(pixel_pin, self.rows * self.columns,
                                          brightness=brightness, auto_write=False,
                                          pixel_order=neopixel.GRB)
-
-    def fill(self, color=0):
-        """
-        Fills all of the NeoPixels with a color or unlit if empty.
-
-        :param color: (Optional) The text or number to display (default=0)
-        :type color: list/tuple or int
-
-        This example shows various ways of using the fill() function
-
-        .. code-block:: python
-
-            import time
-            from adafruit_featherwing import neopixel_featherwing
-
-            neopixel = neopixel_featherwing.NeoPixelFeatherWing()
-            neopixel.fill((255, 255, 255)) # Fill White
-            time.sleep(1)
-            neopixel.fill(0xFF0000) # Fill Red
-            time.sleep(1)
-            neopixel.fill() # Clear all lit NeoPixels
-
-        """
-        super()._fill(color)
-
-    def show(self):
-        """
-        Update the NeoPixels. This is only needed if auto_write is set to False
-        This can be very useful for more advanced graphics effects.
-
-        This example changes the blink rate and prints out the current setting
-
-        .. code-block:: python
-
-            import time
-            from adafruit_featherwing import neopixel_featherwing
-
-            neopixel = neopixel_featherwing.NeoPixelFeatherWing()
-            neopixel.fill() # Clear any lit NeoPixels
-            neopixel.auto_write = False
-            neopixel[0, 0] = (255, 255, 255) # Set White
-            time.sleep(1)
-            neopixel.show() # Update the NeoPixels
-
-        """
-        super()._show()
-
-    def shift_right(self, rotate=False):
-        """
-        Shift all pixels left
-
-        :param rotate: (Optional) Rotate the shifted pixels to the left side (default=False)
-
-        This example shifts 2 pixels to the right
-
-        .. code-block:: python
-
-            import time
-            from adafruit_featherwing import neopixel_featherwing
-
-            neopixel = neopixel_featherwing.NeoPixelFeatherWing()
-
-            # Draw Red and Green Pixels
-            neopixel[4, 1] = (255, 0, 0)
-            neopixel[5, 1] = (0, 255, 0)
-
-            # Rotate it off the screen
-            for i in range(0, neopixel.columns - 1):
-                neopixel.shift_right(True)
-                time.sleep(.1)
-
-            time.sleep(1)
-            # Shift it off the screen
-            for i in range(0, neopixel.columns - 1):
-                neopixel.shift_right()
-                time.sleep(.1)
-
-        """
-        super()._shift_right(rotate)
-
-    def shift_left(self, rotate=False):
-        """
-        Shift all pixels left
-
-        :param rotate: (Optional) Rotate the shifted pixels to the right side (default=False)
-
-        This example shifts 2 pixels to the left
-
-        .. code-block:: python
-
-            import time
-            from adafruit_featherwing import neopixel_featherwing
-
-            neopixel = neopixel_featherwing.NeoPixelFeatherWing()
-
-            # Draw Red and Green Pixels
-            neopixel[4, 1] = (255, 0, 0)
-            neopixel[5, 1] = (0, 255, 0)
-
-            # Rotate it off the screen
-            for i in range(0, neopixel.columns - 1):
-                neopixel.shift_left(True)
-                time.sleep(.1)
-
-            time.sleep(1)
-            # Shift it off the screen
-            for i in range(0, neopixel.columns - 1):
-                neopixel.shift_left()
-                time.sleep(.1)
-
-        """
-        super()._shift_left(rotate)
 
     def shift_up(self, rotate=False):
         """
@@ -195,7 +83,7 @@ class NeoPixelFeatherWing(PixelMatrixFeatherWing):
                 time.sleep(.1)
 
         """
-        super()._shift_down(rotate) # Up and down are reversed
+        super().shift_down(rotate) # Up and down are reversed
 
     def shift_down(self, rotate=False):
         """
@@ -228,64 +116,4 @@ class NeoPixelFeatherWing(PixelMatrixFeatherWing):
                 time.sleep(.1)
 
         """
-        super()._shift_up(rotate) # Up and down are reversed
-
-    @property
-    def auto_write(self):
-        """
-        Whether or not we are automatically updating
-        If set to false, be sure to call show() to update
-
-        This lights NeoPixels with and without auto_write
-
-        .. code-block:: python
-
-            import time
-            from adafruit_featherwing import neopixel_featherwing
-
-            neopixel = neopixel_featherwing.NeoPixelFeatherWing()
-            neopixel.fill() # Clear any lit NeoPixels
-            neopixel[0, 0] = (255, 255, 255) # Set White
-            time.sleep(1)
-
-            neopixel.auto_write = False
-            neopixel[1, 0] = (255, 255, 255) # Set White
-            time.sleep(1)
-            neopixel.show() # Update the NeoPixels
-
-        """
-        return self._auto_write
-
-    @auto_write.setter
-    def auto_write(self, write):
-        if isinstance(write, bool):
-            self._auto_write = write
-
-    @property
-    def brightness(self):
-        """
-        Overall brightness of the display
-
-        This example changes the brightness
-
-        .. code-block:: python
-
-            import time
-            from adafruit_featherwing import neopixel_featherwing
-
-            neopixel = neopixel_featherwing.NeoPixelFeatherWing()
-            neopixel.brightness = 0
-            neopixel.fill(0xFFFFFF)
-            for i in range(0, 6):
-                neopixel.brightness = (i / 10)
-                time.sleep(.2)
-
-            neopixel.brightness = 0.3
-
-        """
-        return self._matrix.brightness
-
-    @brightness.setter
-    def brightness(self, brightness):
-        self._matrix.brightness = min(max(brightness, 0.0), 1.0)
-        self._update()
+        super().shift_up(rotate) # Up and down are reversed
